requirements.txt
streamlit==1.33.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
cryptography==42.0.5
openai
google-generativeai
deepl
beautifulsoup4
nltk
pandas
tweepy==4.14.0